server/flask/requirements.txt
pdf2image==1.16.2
pytesseract==0.3.10
python_dateutil==2.8.2
spacy==3.4.4
Flask
Flask-cors
python-dotenv
